+++ sts/TigerWhale/src/main/java/com/tigerWhale/controller/detailBoardController.java
@@ -12,32 +12,22 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 
 @Controller
-@RequestMapping("/mypage")
+@RequestMapping("/detailBoard")
 public class detailBoardController {
 	
-	@RequestMapping("/mypage")
-	public String myPage() {
-		return "mypage/mypage";
+	@RequestMapping("/detailPage")
+	public String detailPage() {
+		return "detailBoard/detailPage";
 	}
 	
-	@RequestMapping("/mypage-email")
-	public String mypageEmail() {
-		return "mypage/mypage-email";
+	@RequestMapping("/detailBuy")
+	public String detailBuy() {
+		return "detailBoard/detailBuy";
 	}
 	
-	@RequestMapping("/mypage-name")
-	public String mypageName() {
-		return "mypage/mypage-name";
-	}
-	
-	@RequestMapping("/mypage-password")
-	public String mypagePassword() {
-		return "mypage/mypage-password";
-	}
-	
-	@RequestMapping("/mypage-phone")
-	public String mypagePhone() {
-		return "mypage/mypage-phone";
+	@RequestMapping("/detailWrite")
+	public String detailWrite() {
+		return "detailBoard/detailWrite";
 	}
