+++ sts/TigerWhale/src/main/webapp/WEB-INF/detailBoard/detailPage.jsp
@@ -1,9 +1,6 @@
 <%@ page language="java" contentType="text/html; charset=UTF-8"
     pageEncoding="UTF-8"%>
-<!DOCTYPE html>
-<html>
-    <body>
-        <section>
+<section>
             <div class="container CCC">
                 <!--container => 중앙위치 /      container-fluid -->
                 <div class="row">
@@ -487,6 +484,3 @@
 
             var map = new kakao.maps.Map(container, options);
         </script>
-    </body>
-
-</html>
