+++ sts/TigerWhale/src/main/webapp/WEB-INF/detailBoard/detailWrite.jsp
@@ -1,74 +1,6 @@
 <%@ page language="java" contentType="text/html; charset=UTF-8"
     pageEncoding="UTF-8"%>
-<!DOCTYPE html>
-<html>
-    <head>
-        <meta charset="UTF-8">
-        <meta http-equiv="X-UA-Compatible" content="IE=edge">
-        <meta name="viewport" content="width=device-width, initial-scale=1.0">
-        <title>Document</title>
-        <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.4.1/css/bootstrap.min.css">
-        <link rel="stylesheet" href="/publishing/TestWeb/resources/bootstrap/css/bootstrap.css">
-        <link rel="stylesheet" href="/publishing/TestWeb/resources/bootstrap/css/bootstrap-theme.css">
-        <link rel="stylesheet" href="/publishing/TestWeb/resources/bootstrap/css/bootstrap.min.css">
-        <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.5.1/jquery.min.js"></script>
-        <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.4.1/js/bootstrap.min.js"></script>
-        <style>
-            .titlebox {
-                width: 250px;
-                text-align: center;
-                height: 40px;
-                line-height: 40px;
-                margin: 70px auto;
-                border-top: 1px solid #E1E1E1;
-                border-bottom: 1px solid #E1E1E1;
-            }
 
-            .t-control .t-title {
-                width: 100px;
-                min-width: 100px;
-                /*화면줄었을때 최소크기*/
-                vertical-align: middle;
-                /*테이블에서 세로 가운대정렬*/
-            }
-
-            /*input 요소의 넓이만 선택*/
-            .t-control input {
-                width: 200px;
-                display: inline;
-                /*인라인요소로 변경한다*/
-            }
-
-            /*개발자 도구에서 input-sm의 정렬을 푼다*/
-            .t-control .input-sm {
-                height: auto;
-                line-height: normal;
-            }
-
-            /*td의 아래만 밑줄*/
-            .t-control td {
-                border-bottom: 1px solid #E1E1E1;
-            }
-
-            .titlefoot {
-                text-align: center;
-                margin: 30px auto;
-                /*박스의 가운데정렬*/
-                width: 50%;
-                /*크기 50%*/
-            }
-
-            .option,
-            .btn_option {
-                padding: 10px;
-                margin-bottom: 10px;
-                border: 1px solid black;
-            }
-
-        </style>
-    </head>
-
-    <body>
         <section>
             <div class="container">
                 <div class="row">
@@ -190,5 +122,3 @@
             }
         </script>
     </body>
-
-</html>
