+++ sts/TigerWhale/src/main/webapp/WEB-INF/faqBoard/faqList.jsp
@@ -1,110 +1,6 @@
 <%@ page language="java" contentType="text/html; charset=UTF-8"
     pageEncoding="UTF-8"%>
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <link rel="stylesheet" href="/publishing/TestWeb/resources/bootstrap/css/bootstrap.css">
-    <link rel="stylesheet" href="/publishing/TestWeb/resources/bootstrap/css/bootstrap-theme.css">
-    <link rel="stylesheet" href="/publishing/TestWeb/resources/bootstrap/css/bootstrap.min.css">
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Document</title>
-    <style>
-        a {
-            text-decoration: none;
-        }
-
-        a:hover {
-            /* text-decoration: none; */
-        }
-
-        * {
-            margin: 0;
-            padding: 0;
-        }
-
-        ul li {
-            list-style: none;
-        }
-
-        .titlebox {
-            width: 250px;
-            text-align: center;
-            height: 40px;
-            line-height: 40px;
-            margin: 70px auto;
-            border-top: 1px solid #E1E1E1;
-            border-bottom: 1px solid #E1E1E1;
-        }
-
-        /*자게 그리드*/
-        .board-table {
-            float: none;
-            /*col의 플로팅 해제 */
-            margin: 0 auto;
-        }
-
-        /*---검색버튼--- */
-        .search-wrap {
-            overflow: hidden;
-            /*부모요소에 히든 */
-            margin-bottom: 20px;
-        }
-
-        .search-btn,
-        .search-input,
-        .search-select {
-            float: right;
-            /*우측 플로팅 */
-        }
-
-        .search-input {
-            width: 140px;
-        }
-
-        .search-select {
-            width: 80px;
-        }
-
-        /*제목에 클래스를 줘서 강제로 크기 고정*/
-        .board-title {
-            width: 300px;
-        }
-
-        /*th의 순차적 접근 색상변경 */
-        .table-bordered>thead>tr>th {
-            background-color: #E7F2F9;
-            color: #003366;
-        }
-
-        /*테이블 안에 a태그에 순차적 접근 */
-        .table-bordered>tbody>tr>td>a,
-        .table-bordered>tbody>tr>td>a:hover {
-            color: black;
-        }
-
-        .pagination {
-            margin: 0;
-            /*페이지 네이션의 기본 마진 0*/
-        }
-
-        .text-center .btn {
-            float: right;
-            /*오른쪽으로 쌓아서 우측정렬 시킨다*/
-            border: 0;
-            /*border스타일 해제*/
-            border-radius: 0;
-            /*윤곽 0*/
-            padding: 5px 10px;
-            /*버튼 패딩을 페이지네이션과 맞춘다*/
-        }
-    </style>
-</head>
-
-<body>
-    <div class="container">
+<div class="container">
         <div class="row">
             <div class="col-sm-12">
                 <section>
@@ -225,8 +121,3 @@
             </div>
         </div>
     </div>
-
-
-</body>
-
-</html>
