--- sts/TigerWhale/src/main/webapp/WEB-INF/tiles/header.html
@@ -1,62 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Document</title>
-
-     <!-- 부트스트랩 css파일참조 -->
-     <!-- <link rel="stylesheet" href="css/bootstrap.css"> -->
-     <link rel="stylesheet" href="/publishing/TestWeb/bootstrap/css/bootstrap.css">
-     <!-- 부트스트랩 테마파일참조 -->
-     <!-- <link rel="stylesheet" href="css/bootstrap-theme.css"> -->
-     <link rel="stylesheet" href="/publishing/TestWeb/bootstrap/css/bootstrap-theme.css">
-
-     <!-- 부트스트랩 js참조 -->
-     <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.15.3/css/all.css" integrity="sha384-SZXxX4whJ79/gErwcOYf+zWLeJdY/qpuqC4cAa9rOGUstPomtqpuNWT9wdPEn2fk" crossorigin="anonymous">
-     <!-- <script src="js/bootstrap.js"></script> -->
-     <script src="/publishing/TestWeb/resources/bootstrap/js/bootstrap.js"></script>
-    
-     <link rel="stylesheet" href="/publishing/TestWeb/resources/css/header.css ">
-</head>
-<body>
-    
-    <section>
-
-        <div class="container">
-            <div class="row">
-                <div class="col-sm-12" style="float: none; margin:0 auto; padding: 0; background-color: aquamarine;">
-                    <div class="header">
-                        <div>
-                            <img src="/publishing/TestWeb/resources/img/mainPageImg/icon_profile.png">
-                        </div>
-                        <div class="search-tag">
-                            <form>
-                                <input type="text"><a href="#"><a href="#"><i class="fas fa-search fa-2x"></i></a>
-                            </form>
-                        </div>
-                        <div class="icon-set">
-                            <span><a href="#"><i class="far fa-user fa-3x"></i></a></span>
-                            <span><a href="#"><i class="far fa-clone fa-3x"></i></a></span>
-                            <span><a href="#"><i class="fas fa-shopping-basket fa-3x"></i></a></span>
-                        </div>
-                    </div>
-                </div>
-                <div class="col-sm-12 header2">
-                        <div class="category">
-                            <i class="fas fa-bars"></i>  
-                            <span>전체 카테고리</span>
-                        </div>
-                        <div class="regist">
-                            <a href="#">로그인</a>
-                            <a href="#">회원가입</a>
-                        </div>
-                </div>
-            </div>
-        </div>
-
-    </section>
-
-</body>
-</html>
