+++ sts/TigerWhale/src/main/webapp/WEB-INF/tiles/templateMypage.jsp
@@ -1,4 +1,4 @@
-detailBoard/detailBuy<%@ page language="java" contentType="text/html; charset=UTF-8" pageEncoding="UTF-8"%>
+<%@ page language="java" contentType="text/html; charset=UTF-8" pageEncoding="UTF-8"%>
 <%@ taglib uri="http://tiles.apache.org/tags-tiles" prefix="tiles"%>
 <!DOCTYPE html>
 <html>
