+++ sts/TigerWhale/src/main/webapp/WEB-INF/tiles/tiles.xml
@@ -32,15 +32,22 @@
 		<put-attribute name="body" value="/WEB-INF/detailBoard/{1}.jsp" />
 	</definition>
 	
+	<!-- fqa타일즈 테믈릿 -->	
+	<definition name="templateFaqBoard" template="/WEB-INF/tiles/templateFaqBoard.jsp">
+		<put-attribute name="body" value="" />
+	</definition>
+	
+	<definition name="faqBoard/*" extends="templateFaqBoard">
+		<put-attribute name="body" value="/WEB-INF/faqBoard/{1}.jsp" />
+	</definition>
+	
+	
 	
 	
 	
 	 <!--  공통화면에 대한 정의 맵핑-->
 	<definition name="templateBasic" template="/WEB-INF/tiles/templateBasic.jsp">
 		<put-attribute name="header" value="/WEB-INF/tiles/header.jsp" />
-		<put-attribute name="left" value="/WEB-INF/tiles/left.jsp" />
-		<put-attribute name="body" value="" />
-		<put-attribute name="footer"  value="/WEB-INF/tiles/footer.jsp"/>
 	</definition>
 	
 	<definition name="*" extends="templateBasic">
