+++ sts/TigerWhale/src/main/webapp/WEB-INF/views/index.jsp
@@ -18,43 +18,8 @@
     <link rel="stylesheet" href="${pageContext.request.contextPath }/resources/css/main-page.css">
     <link rel="stylesheet" href="${pageContext.request.contextPath }/resources/css/header.css ">
 </head>
-<body>
-
-    <section>
-
-        <div class="container">
-            <div class="row">
-                <div class="col-sm-12" style="float: none; margin:0 auto; padding: 0; background-color: aquamarine;">
-                    <div class="header">
-                        <div>
-                            <img src="${pageContext.request.contextPath }/resources/img/mainPageImg/icon_profile.png">
-                        </div>
-                        <div class="search-tag">
-                            <form>
-                                <input type="text"><a href="#"><a href="#"><i class="fas fa-search fa-2x"></i></a>
-                            </form>
-                        </div>
-                        <div class="icon-set">
-                            <span><a href="#"><i class="far fa-user fa-3x"></i></a></span>
-                            <span><a href="#"><i class="far fa-clone fa-3x"></i></a></span>
-                            <span><a href="#"><i class="fas fa-shopping-basket fa-3x"></i></a></span>
-                        </div>
-                    </div>
-                </div>
-                <div class="col-sm-12 header2">
-                        <div class="category">
-                            <i class="fas fa-bars"></i>  
-                            <span>전체 카테고리</span>
-                        </div>
-                        <div class="regist">
-                            <a href="#">로그인</a>
-                            <a href="#">회원가입</a>
-                        </div>
-                </div>
-            </div>
-        </div>
-
-    </section>
+	
+	<section>
 
     <div class="container main1">
         <div class="row">
@@ -504,7 +469,4 @@
         
     </div>
 
-
-
-</body>
-</html>
+</section>
